+++ src/main/java/com/zhiliao/hotel/service/impl/ZlRepairServiceImpl.java
@@ -3,17 +3,22 @@ package com.zhiliao.hotel.service.impl;
 import com.alibaba.fastjson.JSON;
 import com.google.gson.Gson;
 import com.zhiliao.hotel.common.constant.RedisKeyConstant;
+import com.zhiliao.hotel.common.exception.BizException;
 import com.zhiliao.hotel.controller.Repair.params.RepairParam;
 import com.zhiliao.hotel.controller.Repair.vo.RepairOrderToPhpVO;
 import com.zhiliao.hotel.controller.Repair.vo.RepairOrderVO;
+import com.zhiliao.hotel.controller.myAppointment.dto.ZlRepairorderDTO;
 import com.zhiliao.hotel.controller.myOrder.vo.OrderPhpSendVO;
 import com.zhiliao.hotel.controller.myOrder.vo.OrderPhpVO;
 import com.zhiliao.hotel.mapper.ZlHotelRoomMapper;
 import com.zhiliao.hotel.mapper.ZlRepairorderMapper;
 import com.zhiliao.hotel.mapper.ZlWxuserMapper;
+import com.zhiliao.hotel.mapper.ZlWxuserdetailMapper;
 import com.zhiliao.hotel.model.ZlHotelroom;
 import com.zhiliao.hotel.model.ZlRepairorder;
 import com.zhiliao.hotel.model.ZlWxuser;
+import com.zhiliao.hotel.model.ZlWxuserdetail;
+import com.zhiliao.hotel.service.OrderLogService;
 import com.zhiliao.hotel.service.ZlRepairService;
 import com.zhiliao.hotel.utils.*;
 import org.apache.commons.lang3.StringUtils;
@@ -49,35 +54,29 @@ public class ZlRepairServiceImpl implements ZlRepairService {
     @Autowired
     private ZlWxuserMapper zlWxuserMapper;
 
-    @Override
-    public void addRepairMsg(ZlRepairorder repair) {
-//        zlRepairorderMapper.insertRepairorder(repair);
-        zlRepairorderMapper.insertSelective(repair);
-    }
+    @Autowired
+    private OrderLogService orderLogService;
+
+    @Autowired
+    private ZlWxuserdetailMapper zlWxuserdetailMapper;
 
     @Override
-    public Map<String, Object> addRepairMsg(Long userid, RepairParam repairParam, MultipartFile multipartFile) {
+    public Map<String, Object> addRepairMsg(Long userid, RepairParam repairParam) {
+
+        ZlWxuserdetail zlWxuserdetail = new ZlWxuserdetail();
+        zlWxuserdetail.setUserid(userid);
+        zlWxuserdetail = zlWxuserdetailMapper.selectOne(zlWxuserdetail);
+
         Map<String, Object> map = new HashMap<>();
 
-        //说明不是第一次
-        if (repairParam.getOrderid() != -1) {
-            ZlRepairorder zlRepairorder = new ZlRepairorder();
-            zlRepairorder.setOrderid(repairParam.getOrderid());
-            List<ZlRepairorder> zlRepairorderList = zlRepairorderMapper.select(zlRepairorder);
-            ZlRepairorder zlRepairorderOld = zlRepairorderList.get(0);
-            StringBuffer imageurls = new StringBuffer();
-            String imageurlOld = zlRepairorderOld.getImgurls();
-            String imgurl = UploadPhotoUtil.uploadPhotoUtil2(multipartFile);
-            imageurls.append(imageurlOld).append("|").append(imgurl);
-            zlRepairorderMapper.updateCommentImg(repairParam.getOrderid(), imageurls.toString());
-            map.put("orderid", repairParam.getOrderid());
-            return map;
-        }
-
         Integer now = DateUtils.javaToPhpNowDateTime();
-        String serialNumber = OrderIDUtil.createOrderID("");
+        String serialNumber = OrderIDUtil.createOrderID("WX");
         ZlRepairorder zlRepairorder = new ZlRepairorder();
         zlRepairorder.setUserid(userid);
+        zlRepairorder.setUsername(zlWxuserdetail.getRealname());
+        if (StringUtils.isNoneBlank(zlWxuserdetail.getTel())) {
+            zlRepairorder.setTel(zlWxuserdetail.getTel());
+        }
         zlRepairorder.setHotelname(repairParam.getHotelname());  // 酒店名
         zlRepairorder.setHotelid(repairParam.getHotelid());     // 酒店ID
         zlRepairorder.setRoomid(repairParam.getRoomid());       // 客房ID
@@ -86,19 +85,21 @@ public class ZlRepairServiceImpl implements ZlRepairService {
         zlRepairorder.setSerialnumber(serialNumber);  //订单号
         zlRepairorder.setCreatedate(now);
         zlRepairorder.setUpdatedate(now);
+        zlRepairorder.setImgurls(repairParam.getImgurls());
 
         ZlHotelroom zlHotelroom = zlHotelRoomMapper.getByHotelIDAndRoomNumber(repairParam.getRoomnumber(), repairParam.getHotelid());
+        if (zlHotelroom == null) {
+            throw new BizException("您的码是前台码，不提供该服务");
+        }
         zlRepairorder.setFloornumber(zlHotelroom.getRoomfloor());
         ZlWxuser zlWxuser = new ZlWxuser();
         zlWxuser.setUserid(userid);
         ZlWxuser wxuser = zlWxuserMapper.selectOne(zlWxuser);
         zlRepairorder.setUsername(wxuser.getNickname());
-
-        if (multipartFile != null) {
-            String imgurl = UploadPhotoUtil.uploadPhotoUtil2(multipartFile);
-            zlRepairorder.setImgurls(imgurl);  //图片地址
-        }
         zlRepairorderMapper.insertSelective(zlRepairorder);
+
+        map.put("orderid", zlRepairorder.getOrderid());
+        map.put("serialnumber", zlRepairorder.getSerialnumber());
 
         logger.info("报修订单插入数据库完成,订单id:" + zlRepairorder.getOrderid());
         // 推送消息
@@ -113,17 +114,16 @@ public class ZlRepairServiceImpl implements ZlRepairService {
         stringRedisTemplate.convertAndSend(RedisKeyConstant.TOPIC_REPAIR_ORDER, orderStr);
         logger.info("推送报修订单到redis通知php后台人员完成,订单信息:" + repairOrderToPhpVO);
 
-        map.put("orderid", zlRepairorder.getOrderid());
         return map;
     }
 
 
     @Override
-    public RepairOrderVO findRepairOrder(Long orderID) {
-        RepairOrderVO repairOrderVO = zlRepairorderMapper.queryRepairMsg(orderID);
+    public ZlRepairorderDTO findRepairOrder(Long orderID) {
+        ZlRepairorderDTO zlRepairorderDTO = zlRepairorderMapper.queryRepairMsg(orderID);
         List<String> imageurllist = new LinkedList<>();
-        if (StringUtils.isNoneBlank(repairOrderVO.getImgurls())) {
-            String imageurls = repairOrderVO.getImgurls();
+        if (StringUtils.isNoneBlank(zlRepairorderDTO.getImgurls())) {
+            String imageurls = zlRepairorderDTO.getImgurls();
             if (imageurls.contains("|")) {
                 String[] imageurlArr = imageurls.split("\\|");
                 for (int i = 0; i < imageurlArr.length; i++) {
@@ -133,12 +133,24 @@ public class ZlRepairServiceImpl implements ZlRepairService {
                 imageurllist.add(imageurls);
             }
         }
-        repairOrderVO.setImageurllist(imageurllist);
-        return repairOrderVO;
+        zlRepairorderDTO.setImageurllist(imageurllist);
+        return zlRepairorderDTO;
     }
 
     @Override
-    public void cancelRepairOrder(Long orderID, Integer updatedate) {
+    public void cancelRepairOrder(Long orderID) {
+
+        Integer updatedate = DateUtils.javaToPhpNowDateTime();
+
+        ZlRepairorder zlRepairorder = new ZlRepairorder();
+        zlRepairorder.setOrderid(orderID);
+        zlRepairorder = zlRepairorderMapper.selectOne(zlRepairorder);
+        if (zlRepairorder == null) {
+            throw new BizException("参数有误!");
+        }
+
+        //将订单取消操作写到记录表中
+        orderLogService.cancelOrderLog(zlRepairorder.getOrderid(), zlRepairorder.getHotelid(), zlRepairorder.getCreatedate(), zlRepairorder.getMoldtype());
         zlRepairorderMapper.removeRepairOrder(orderID, updatedate);
     }
